+++ cobaya/run.py
@@ -52,7 +52,7 @@ def run(info_or_yaml_or_file: Union[InputDict, str, os.PathLike],
     :param test: only test initialization rather than actually running
     :param override: option dictionary to merge into the input one, overriding settings
        (but with lower precedence than the explicit keyword arguments)
-    :param allow_changes: if true, allow input option changes when resuming
+    :param allow_changes: if true, allow input option changes when resuming or minimizing
     :return: (updated_info, sampler) tuple of options dictionary and Sampler instance,
               or (updated_info, post_results) if using "post" post-processing
     """
@@ -181,8 +181,9 @@ def run_script(args=None):
                               "process (incompatible with post-processing)."),
                         **trueNone_kwargs)
     parser.add_argument("--allow-changes",
-                        help="Set to allow changing parameters parameters when resuming",
-                        action="store_true", default=True)
+                        help="Allow changing input parameters when resuming "
+                             "or minimizing, skipping consistency checks",
+                        **trueNone_kwargs)
     parser.add_argument("--version", action="version", version=get_version())
     parser.add_argument("--no-mpi",
                         help=("disable MPI when mpi4py installed but MPI does "